feat(chat): add journal persistence and isolate chat workflow
Add v2 session journals, idempotent sends, resume bindings, run logs, exports, and graceful process shutdown. Separate Chat from Task and Planning subscriptions, index active event replay, and fix debug process matching.

Validated: 172 frontend tests, 289 Rust unit tests, 2 integration tests, desktop builds, and native navigation/export/restart checks. Real CLI permission and stop acceptance remains pending authorization.

+++ src-tauri/src/run_recovery.rs
@@ -271,7 +271,7 @@ mod tests {
                 "run-running",
                 &task.id,
                 ProcessKind::Command,
-                u32::MAX - 1,
+                (i32::MAX as u32) - 1,
                 None,
             ))
             .unwrap();
@@ -280,7 +280,7 @@ mod tests {
                 "review-active",
                 &task.id,
                 ProcessKind::ImplementationReview,
-                u32::MAX,
+                i32::MAX as u32,
                 None,
             ))
             .unwrap();
